requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pyTelegramBotAPI==4.15.4
python-decouple==3.8
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.2.0
wikipedia==1.4.0
